[REFACTOR][SCRIPT] Simplify expression string docs
